Another engine fix

+++ src/main/java/org/turing/app/controllers/ProgramEditController.java
@@ -4,10 +4,12 @@ import org.turing.app.common.State;
 import org.turing.app.common.Symbol;
 import org.turing.app.exceptions.SymbolException;
 import org.turing.app.model.ActionTriple;
+import org.turing.app.model.DataModel;
 import org.turing.app.model.ProgramModel;
 import org.turing.app.views.constants.ApplicationStrings;
 import org.turing.app.views.panels.StatePanel;
 import org.turing.app.views.panels.TapePanel;
+import sun.jvm.hotspot.debugger.cdbg.Sym;
 
 import javax.swing.*;
 import javax.swing.table.AbstractTableModel;
@@ -20,14 +22,16 @@ public class ProgramEditController {
     private static final ActionTriple defaultActionTriple = new ActionTriple(HALT, BLANK, NONE);
 
     private final ProgramModel programModel;
+    private final DataModel dataModel;
 
     private JFrame programFrame;
     private JTable programTable;
     private StatePanel statePanel;
     private TapePanel tapePanel;
 
-    public ProgramEditController(ProgramModel programModel) {
+    public ProgramEditController(ProgramModel programModel, DataModel dataModel) {
         this.programModel = programModel;
+        this.dataModel = dataModel;
     }
 
     public void onStateAddition() {
@@ -53,8 +57,16 @@ public class ProgramEditController {
 
     public void onStateDeletion() {
         try {
-            if (programTable.getSelectedRow() >= 0)
-                programModel.deleteState(programModel.getStateAt(programTable.getSelectedRow()));
+            if (programTable.getSelectedRow() >= 0) {
+                if(dataModel.getState() != programModel.getStateAt(programTable.getSelectedRow()))
+                    programModel.deleteState(programModel.getStateAt(programTable.getSelectedRow()));
+                else
+                {
+                    JOptionPane.showMessageDialog(
+                            programFrame, "Cannot delete state " + dataModel.getState().getName() + " it is current state.", "Error", JOptionPane.ERROR_MESSAGE);
+                    return;
+                }
+            }
         } catch (RuntimeException ex) {
             JOptionPane.showMessageDialog(
                     programFrame, ex.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
@@ -98,8 +110,16 @@ public class ProgramEditController {
 
     public void onSymbolDeletion() {
         try {
-            if (programTable.getSelectedColumn() > 0)
-                programModel.deleteSymbol(programModel.getSymbolAt(programTable.getSelectedColumn() - 1));
+            if (programTable.getSelectedColumn() > 0) {
+                if(canSymbolBeRemoved(programModel.getSymbolAt(programTable.getSelectedColumn() - 1)))
+                    programModel.deleteSymbol(programModel.getSymbolAt(programTable.getSelectedColumn() - 1));
+                else
+                {
+                    JOptionPane.showMessageDialog(
+                            programFrame, "Cannot delete symbol " + dataModel.read().getValue() + ". It is still used on a tape.", "Error", JOptionPane.ERROR_MESSAGE);
+                    return;
+                }
+            }
         } catch (RuntimeException ex) {
             JOptionPane.showMessageDialog(
                     programFrame, ex.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
@@ -107,6 +127,15 @@ public class ProgramEditController {
         }
 
         ((AbstractTableModel) programTable.getModel()).fireTableStructureChanged();
+    }
+
+    private boolean canSymbolBeRemoved(Symbol symbol)
+    {
+        boolean canBeRemoved = true;
+        canBeRemoved &= !symbol.getValue().equalsIgnoreCase(dataModel.read().getValue());
+        canBeRemoved &= !dataModel.getLeftTape().contains(symbol);
+        canBeRemoved &= !dataModel.getRightTape().contains(symbol);
+        return canBeRemoved;
     }
 
     public ProgramModel getProgramModel() {
